% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterComputation}
\alias{clusterComputation}
\title{\strong{clusterComputation}}
\usage{
\code{clusterComputation()}
}
\arguments{
\item{No}{additional arguments are needed}
}
\description{
clusterComputation is a function to cluster the different distances from the \code{distance} function together.
To speed up the code the function is using parallel computation.
The function first counts the number of columns of the distanceMatrixDF to determine the number of repetitions.
It then determine the number of cores in the computer and builds a cluster on each core with respect to one free core.
After this the function exports \code{cluster}, \code{distanceMatrixDF}, \code{numberOfColumns}, \code{emptyClassMeteroites}
and \code{classMeteroites} to each core.
The function uses parSapply to parallel the computation.
The function returns the \code{classDF} dat frame with the classes and the number of meteroites in each category and stops automatically the cluster.
}
