% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{distance}
\alias{distance}
\title{\strong{distance}}
\usage{
\code{distance()}
}
\arguments{
\item{No}{additional arguments are used}
}
\description{
The distance function computes the string distances between the classes column of the meteroitesData data frame and pre defined meteroite types.
For measuring the distance between to strings the Jaro-Winker distance is used.
A value of 1 means there is no similarity between two strings.
Vice versa a value of 0 means the similarity between two strings is 100%

The function itself creates four variable

\code{classMeteroites} contains the general classes of different types of meteroites.
\code{vectorClass} is a vector containing charackters, sliced out of the meteroitesData file.
\code{distanceMatrix} a matrix ccrated by the stringdistmatrix function
\code{distanceMatrixDF} a data frame created from the distanceMatrix variable
}
